pyproject.toml
[project]
name = "cs5224-ai-renovator"
version = "0.1.0"
description = "Backend API for the AI Renovator interior design app"
requires-python = "==3.12.*"

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "ikea-api[httpx,wrappers]>=2.4.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-core>=0.3.0",
    "mangum>=0.21.0",
    "sqlalchemy>=2.0.49",
    "psycopg2-binary>=2.9.11",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "boto3>=1.34.0",
]

[tool.uv]
dev-dependencies = []
